+++ R/sim_geno.R
@@ -76,8 +76,10 @@ function(cross, map=NULL, n_draws=1, error_prob=1e-4,
     }
 
     # pseudomarker map
-    if(is.null(map))
+    if(is.null(map)) {
+        if(is.null(cross$gmap)) stop("If cross does not contain a genetic map, map must be provided.")
         map <- insert_pseudomarkers(cross$gmap)
+    }
     # possibly subset the map
     if(length(map) != length(cross$geno) || !all(names(map) == names(cross$geno))) {
         chr <- names(cross$geno)
@@ -139,7 +141,7 @@ function(cross, map=NULL, n_draws=1, error_prob=1e-4,
 
     }
 
-    names(draws) <- names(cross$gmap)
+    names(draws) <- names(cross$geno)
     attr(draws, "crosstype") <- cross$crosstype
     attr(draws, "is_x_chr") <- cross$is_x_chr
     attr(draws, "alleles") <- cross$alleles